+++ django_api/django_api/apps/ocha/imports/utilities.py
@@ -1,19 +1,16 @@
 import logging
 from time import sleep
 
-import itertools
 import requests
 from requests import RequestException
 from requests.status_codes import codes
 
-from cluster.models import Cluster, ClusterObjective, ClusterActivity
-from core.common import EXTERNAL_DATA_SOURCES, CLUSTER_TYPES
-from core.models import ResponsePlan, Country, GatewayType, Location
+from cluster.models import Cluster, ClusterObjective
+from core.common import EXTERNAL_DATA_SOURCES
+from core.models import Country, GatewayType, Location
 from indicator.models import Reportable, IndicatorBlueprint
-from ocha.constants import HPC_V2_ROOT_URL, HPC_V1_ROOT_URL, RefCode
+from ocha.constants import HPC_V2_ROOT_URL
 from ocha.imports.bulk import fetch_json_urls_async
-from ocha.imports.serializers import V2PartnerProjectImportSerializer, V1FundingSourceImportSerializer, \
-    V1ResponsePlanImportSerializer
 from ocha.utilities import get_dict_from_list_by_key
 
 logger = logging.getLogger('ocha-sync')
@@ -110,84 +107,6 @@ def save_location_list(location_list):
     return locations
 
 
-def import_project_details(project, current_version_id):
-    source_url = HPC_V2_ROOT_URL + 'project-version/{}/attachments'.format(current_version_id)
-    attachments = get_json_from_url(source_url)['data']
-
-    reportables = []
-
-    for attachment in attachments:
-        if attachment['attachment']['type'] == 'indicator':
-            blueprint, _ = IndicatorBlueprint.objects.update_or_create(
-                external_source=EXTERNAL_DATA_SOURCES.HPC,
-                external_id=attachment['attachment']['id'],
-                defaults={
-                    'title': attachment['attachment']['value']['description'],
-                }
-            )
-
-            totals = attachment['attachment']['value']['metrics']['values']['totals']
-            disaggregated = attachment['attachment']['value']['metrics']['values']['disaggregated']
-
-            target = get_dict_from_list_by_key(totals, 'Target', key='name.en')['value']
-            in_need = get_dict_from_list_by_key(totals, 'In Need', key='name.en')['value']
-            baseline = get_dict_from_list_by_key(totals, 'Baseline', key='name.en')['value']
-
-            locations = save_location_list(disaggregated['locations'])
-
-            # TODO: Parent content_object
-            reportable, _ = Reportable.objects.update_or_create(
-                external_source=EXTERNAL_DATA_SOURCES.HPC,
-                external_id=attachment['attachment']['id'],
-                defaults={
-                    'blueprint': blueprint,
-                    'target': target,
-                    'in_need': in_need,
-                    'baseline': baseline,
-                }
-            )
-
-            reportable.locations.add(*locations)
-            reportables.append(reportable)
-
-    logger.debug('Saving {} reportables for {}'.format(
-        len(reportables), project
-    ))
-    project.reportables.add(*reportables)
-
-
-def import_project(external_project_id, response_plan=None):
-    source_url = HPC_V2_ROOT_URL + 'project/{}'.format(external_project_id)
-    project_data = get_json_from_url(source_url)
-    serializer = V2PartnerProjectImportSerializer(data=project_data['data'])
-    serializer.is_valid(raise_exception=True)
-    project = serializer.save()
-
-    from ocha.tasks import finish_partner_project_import
-    finish_partner_project_import.delay(
-        project.pk, response_plan_id=getattr(response_plan, 'id', None)
-    )
-
-    return project
-
-
-def import_response_plan(external_plan_id, workspace=None):
-    logger.debug('Importing Response Plan #{}'.format(external_plan_id))
-    source_url = HPC_V1_ROOT_URL + 'rpm/plan/id/{}?format=json&content=entities'.format(external_plan_id)
-    plan_data = get_json_from_url(source_url)['data']
-    if workspace:
-        plan_data['workspace_id'] = workspace.id
-    plan_serializer = V1ResponsePlanImportSerializer(data=plan_data)
-    plan_serializer.is_valid(raise_exception=True)
-    response_plan = plan_serializer.save()
-
-    # Do most of the work in background, otherwise it times out the request a lot
-    from ocha.tasks import finish_response_plan_import
-    finish_response_plan_import.delay(external_plan_id)
-
-    return response_plan
-
-
 def save_cluster_objective(objective, child_activity):
     cluster_id = objective.get('parentId') or child_activity.get('parentId')
     cluster = Cluster.objects.filter(
@@ -245,78 +164,3 @@ def save_reportables_for_cluster_objective_or_activity(objective_or_activity, at
         )
         reportables.append(reportable)
     objective_or_activity.reportables.add(*reportables)
-
-
-def save_activities_and_objectives_for_response_plan(entities_response={}, measurements_response={}):
-    activities = []
-    objectives = {}
-
-    plan_entity_list = entities_response['planEntities'] + measurements_response['planEntities']
-    for entity in plan_entity_list:
-        if entity['entityPrototype']['refCode'] in {RefCode.CLUSTER_OBJECTIVE, RefCode.STRATEGIC_OBJECTIVE}:
-            objectives[entity['id']] = entity
-        elif entity['entityPrototype']['refCode'] == RefCode.CLUSTER_ACTIVITY:
-            activities.append(entity)
-    logger.debug('Found {} objectives and {} activities'.format(
-        len(objectives), len(activities)
-    ))
-
-    for activity in activities:
-        try:
-            parent_objective_ids = list(itertools.chain(*[
-                s['planEntityIds'] for s in activity['value']['support']
-            ]))
-
-            if len(parent_objective_ids) > 1:
-                logger.warning(
-                    'Activity \n`{}` supports \n{} \nobjectives. Only 1st one will be saved.'.format(
-                        activity['value']['description'],
-                        [objectives[obj_id]['value']['description'] for obj_id in parent_objective_ids]
-                    )
-                )
-            parent_objective = objectives[parent_objective_ids[0]]
-        except (KeyError, IndexError):
-            logger.warning('Activity #{} has no objective info'.format(activity['id']))
-            continue
-
-        cluster_objective = save_cluster_objective(parent_objective, activity)
-        if cluster_objective:
-            cluster_activity, _ = ClusterActivity.objects.update_or_create(
-                external_id=activity['id'],
-                external_source=EXTERNAL_DATA_SOURCES.HPC,
-                defaults={
-                    'cluster_objective': cluster_objective,
-                    'title': activity['value']['description'][:2048]
-                }
-            )
-            save_reportables_for_cluster_objective_or_activity(cluster_activity, activity['attachments'])
-
-
-def get_plan_list_for_country(country_iso3):
-    source_url = HPC_V1_ROOT_URL + 'plan/country/{}'.format(country_iso3)
-    try:
-        return get_json_from_url(source_url)['data']
-    except Exception:
-        logger.exception('Error trying to list plans for country')
-        return []
-
-
-def get_project_list_for_plan(plan_id):
-    source_url = HPC_V1_ROOT_URL + 'project/plan/{}'.format(plan_id)
-    try:
-        return get_json_from_url(source_url)['data']
-    except Exception:
-        logger.exception('Error trying to list projects for response plan')
-        return []
-
-
-def import_plans_for_country(country_iso3):
-    plans = get_plan_list_for_country(country_iso3)
-    logger.debug('Importing {} Response Plans for {}'.format(
-        len(plans), country_iso3
-    ))
-    for plan in plans:
-        try:
-            import_response_plan(plan['id'])
-        except Exception:
-            logger.exception('Problem importing Response Plan #{}'.format(plan['id']))
