+++ django_api/django_api/apps/ocha/management/commands/import_projects.py
@@ -2,7 +2,7 @@
 from django.conf import settings
 from django.core.management.base import BaseCommand
 
-from ocha.imports.utilities import import_project
+from ocha.imports.project import import_project
 
 
 class Command(BaseCommand):
