+++ django_api/django_api/apps/ocha/tasks.py
@@ -7,8 +7,9 @@ from core.common import EXTERNAL_DATA_SOURCES, CLUSTER_TYPES
 from core.models import ResponsePlan
 from ocha.constants import HPC_V1_ROOT_URL, HPC_V2_ROOT_URL
 from ocha.imports.serializers import V1FundingSourceImportSerializer
-from ocha.imports.utilities import get_json_from_url, save_activities_and_objectives_for_response_plan, \
-    import_response_plan, import_project_details
+from ocha.imports.utilities import get_json_from_url
+from ocha.imports.response_plan import import_response_plan, save_activities_and_objectives_for_response_plan
+from ocha.imports.project import import_project_details
 from partner.models import PartnerProject
 
 logger = logging.getLogger('ocha-sync')
